backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
async-timeout==4.0.2
attrs==23.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
billiard==3.6.4.0
boto3==1.18.65
botocore==1.21.65
cachetools==5.3.1
cbor2==5.4.6
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
country-converter==1.0.0
cryptography==37.0.4
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dj-rest-auth==3.0.0
Django==2.2.28
django-allauth==0.51.0
django-bootstrap4==22.1
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-cors-headers==2.2.0
django-environ==0.9.0
django-extensions==3.1.5
django-jazzmin==2.6.0
django-storages==1.12.3
django-timezone-field==5.0
djangorestframework==3.13.1
djangorestframework-gis==1.0
drf-yasg==1.20.3
factory-boy==3.2.1
Faker==18.10.1
filetype==1.2.0
flower==1.2.0
frozenlist==1.3.3
future==0.18.3
geocoder==1.38.1
geographiclib==2.0
geopy==2.3.0
google-api-core==2.10.2
google-auth==2.9.0
google-cloud-core==2.3.2
google-cloud-secret-manager==2.11.1
google-cloud-storage==2.4.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpc-google-iam-v1==0.12.6
grpcio==1.54.2
grpcio-status==1.48.2
humanize==4.6.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
jmespath==0.10.0
kombu==5.3.0
MarkupSafe==2.1.3
mixpanel==4.10.0
multidict==6.0.4
numpy==1.24.3
oauthlib==3.2.2
packaging==23.1
pandas==2.0.2
Pillow==9.2.0
pluggy==1.0.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==3.20.3
psycopg2-binary==2.8.6
pubnub==7.0.1
py==1.11.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycountry==22.3.5
pycparser==2.21
pycryptodomex==3.18.0
PyJWT==2.7.0
pyotp==2.7.0
pytest==7.1.2
python-crontab==2.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
ratelim==0.1.6
redis==4.3.4
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
s3transfer==0.5.2
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
stripe==4.2.0
tomli==2.0.1
tornado==6.3.2
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
vine==5.0.0
waitress==2.1.2
wcwidth==0.2.6
whitenoise==6.0.0
wrapt==1.15.0
yarl==1.9.2
django-dbbackup==4.0.2
sendgrid==6.8.3